Move remaining code-lab specific bits to the codelab config
