+++ exercises/ng2ts/ng2ts.ts
@@ -1,3 +1,50 @@
+
+//  TODO: This should be done using require.context
+const preloadedFiles = {
+  'app.component.ts': require(`!raw!./app.component.ts`),
+  'app.module.ts': require('!raw!./app.module.ts'),
+  'app.html': require('!raw!./app.html'),
+  'main.ts': require('!raw!./main.ts'),
+  'video/video-item.ts': require('!raw!./video/video-item.ts'),
+  'api.service.ts': require('!raw!./api.service.ts'),
+  'video/video.service.ts': require('!raw!./video/video.service.ts'),
+  'video/video.html': require('!raw!./video/video.html'),
+  'video/video.component.ts': require('!raw!./video/video.component.ts'),
+  'thumbs/thumbs.component.ts': require('!raw!./thumbs/thumbs.component.ts'),
+  'thumbs/thumbs.html': require('!raw!./thumbs/thumbs.html'),
+  'toggle-panel/toggle-panel.html': require('!raw!./toggle-panel/toggle-panel.html'),
+  'toggle-panel/toggle-panel.component.ts': require('!raw!./toggle-panel/toggle-panel.component.ts'),
+  'wrapper.component.ts': require('!raw!./wrapper.component.ts'),
+  'context/context.component.ts': require('!raw!./context/context.component.ts'),
+  'context/context.service.ts': require('!raw!./context/context.service.ts'),
+  'context/context.html': require('!raw!./context/context.html'),
+  'typescript-intro/Codelab.ts': require('!raw!./typescript-intro/Codelab.ts'),
+  'typescript-intro/Main.ts': require('!raw!./typescript-intro/Main.ts'),
+  'typescript-intro/Guest.ts': require('!raw!./typescript-intro/Guest.ts'),
+  'fuzzy-pipe/fuzzy.pipe.ts': require('!raw!./fuzzy-pipe/fuzzy.pipe.ts'),
+  'tests/codelabTest.ts': require('!raw!./tests/codelabTest.ts'),
+  'tests/createComponentTest.ts': require('!raw!./tests/createComponentTest.ts'),
+  'tests/createModuleTest.ts': require('!raw!./tests/createModuleTest.ts'),
+  'tests/bootstrapTest.ts': require('!raw!./tests/bootstrapTest.ts'),
+  'tests/templatePageSetupTest.ts': require('!raw!./tests/templatePageSetupTest.ts'),
+  'tests/templateAddActionTest.ts': require('!raw!./tests/templateAddActionTest.ts'),
+  'tests/templateAllVideosTest.ts': require('!raw!./tests/templateAllVideosTest.ts'),
+  'tests/diInjectServiceTest.ts': require('!raw!./tests/diInjectServiceTest.ts'),
+  'tests/videoComponentCreateTest.ts': require('!raw!./tests/videoComponentCreateTest.ts'),
+  'tests/videoComponentUseTest.ts': require('!raw!./tests/videoComponentUseTest.ts'),
+  'tests/ThumbsComponentCreateTest.ts': require('!raw!./tests/ThumbsComponentCreateTest.ts'),
+  'tests/ThumbsComponentUseTest.ts': require('!raw!./tests/ThumbsComponentUseTest.ts'),
+  'tests/togglePanelComponentCreateTest.ts': require('!raw!./tests/togglePanelComponentCreateTest.ts'),
+  'tests/togglePanelComponentUseTest.ts': require('!raw!./tests/togglePanelComponentUseTest.ts'),
+  'tests/contextComponentUseTest.ts': require('!raw!./tests/contextComponentUseTest.ts'),
+  'tests/fuzzyPipeCreateTest.ts': require('!raw!./tests/fuzzyPipeCreateTest.ts'),
+  'tests/fuzzyPipeUseTest.ts': require('!raw!./tests/fuzzyPipeUseTest.ts'),
+  'thumbs.app.module.ts': require('!raw!./thumbs.app.module.ts'),
+  'toggle-panel.app.module.ts': require('!raw!./toggle-panel.app.module.ts'),
+  'index.html': '<my-thumbs></my-thumbs><my-wrapper></my-wrapper>'
+};
+
+
 const files = {
   appComponent: 'app.component.ts',
   appModule: 'app.module.ts',
@@ -107,6 +154,7 @@ const stages: stage[] = [
 export interface CodelabConfigTemplate {
   name: string;
   files: {[key: string]: string},
+  preloadedFiles: {[key: string]: string},
   overrides: {
     file: {[key: string]: {[key: string]: string}},
     stage: {[key: string]: {[key: string]: string}},
@@ -143,7 +191,8 @@ export interface MilestoneConfigTemplate {
 
 export const ng2tsConfig: CodelabConfigTemplate = {
   name: 'Angular2 codelab',
-  files: files,
+  files,
+  preloadedFiles,
   overrides: {
     file: fileOverrides,
     stage: stageOverrides
